join Buli on his jurny
this is a small school project i had a lot of fun making. i hope it will give you joy as well.
this is a command line based game, you can compile it yourself and play it!
 your task is to collect all points (#) before the bear comes for you.
it wasn't tested for macOS or Linux so currently and probably for the end of time it will be only windows.